requirements.txt
click==7.1.2
PyGithub==1.53
requests==2.24.0